requirements.txt
asgiref==3.4.1
certifi==2021.10.8
coverage==6.2
Django==4.0
djangorestframework==3.12.4
numpy==1.21.4
pandas==1.3.4
parameterized==0.7.4
psycopg2-binary==2.9.2
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
